Added Design Proposal
